feat(install): add install instructions, fix release (#8)
* feat(install): add install instructions, fix release


* doc(usage): add usage docs, add readme details


* doc(fix): fix go install path in readme


---------
