--- doc/demos.md
@@ -1,47 +1,0 @@
-# Demos
-
-> Demos of the `image-prune` CLI
-
-## Contents
-
-- [List Tags](#list-tags)
-    - [Default](#default)
-    - [Before version (using tag)](#before-version-using-tag)
-    - [Before version (using label)](#before-version-using-label)
-    - [Before time](#before-time)
-- [Prune](#prune)
-    - [Before version (using tag)](#before-version-using-tag-1)
-    - [Before version (using label)](#before-version-using-label-1)
-    - [Before time](#before-time-1)
-
-## List tags
-
-### Default
-
-![image-prune list-tags](./img/list-tags.gif)
-
-### Before version (using tag)
-
-![image-prune list-tags --before-version](./img/list-tags-before-version-tag.gif)
-
-### Before version (using label)
-
-![image-prune list-tags --before-version --version-label](./img/list-tags-before-version-label.gif)
-
-### Before time
-
-![image-prune list-tags --before-time](./img/list-tags-before-time.gif)
-
-## Prune
-
-### Before version (using tag)
-
-![image-prune prune --before-version](./img/prune-before-version-tag.gif)
-
-### Before version (using label)
-
-![image-prune prune --before-version --version-label](./img/prune-before-version-label.gif)
-
-### Before time
-
-![image-prune prune --before-time](./img/prune-before-time.gif)
